+++ test/eca/llm_providers/ollama_test.clj
@@ -7,28 +7,25 @@
 (deftest ->messages-with-history-test
   (testing "no previous history"
     (is (match?
-         [{:role "system" :content "You are ECA"}]
-         (#'llm-providers.ollama/past-messages->messages [] "You are ECA"))))
+         []
+         (#'llm-providers.ollama/normalize-messages []))))
   (testing "With basic text history"
     (is (match?
-         [{:role "system" :content "You are ECA"}
-          {:role "user" :content "Count with me: 1"}
+         [{:role "user" :content "Count with me: 1"}
           {:role "assistant" :content "2"}]
-         (#'llm-providers.ollama/past-messages->messages
+         (#'llm-providers.ollama/normalize-messages
           [{:role "user" :content "Count with me: 1"}
-           {:role "assistant" :content "2"}]
-          "You are ECA"))))
+           {:role "assistant" :content "2"}]))))
   (testing "With tool_call history"
     (is (match?
-         [{:role "system" :content "You are ECA"}
-          {:role "user" :content "List the files you are allowed"}
+         [{:role "user" :content "List the files you are allowed"}
           {:role "assistant" :content "Ok!"}
           {:role "assistant" :tool-calls [{:type "function"
                                            :function {:name "list_allowed_directories"
                                                       :arguments {}}}]}
           {:role "tool" :content "Allowed directories: /foo/bar\n"}
           {:role "assistant" :content "I see /foo/bar"}]
-         (#'llm-providers.ollama/past-messages->messages
+         (#'llm-providers.ollama/normalize-messages
           [{:role "user" :content "List the files you are allowed"}
            {:role "assistant" :content "Ok!"}
            {:role "tool_call" :content {:id "call-1" :name "list_allowed_directories" :arguments {}}}
@@ -38,5 +35,4 @@
                                                :output {:contents [{:type :text
                                                                     :error false
                                                                     :content "Allowed directories: /foo/bar"}]}}}
-           {:role "assistant" :content "I see /foo/bar"}]
-          "You are ECA")))))
+           {:role "assistant" :content "I see /foo/bar"}])))))